+++ nodeconductor/core/filters.py
@@ -331,8 +331,9 @@ class CategoryFilter(django_filters.CharFilter):
 
 class StaffOrUserFilter(object):
     def filter_queryset(self, request, queryset, view):
-        if request.user.is_staff:
+        if request.user.is_staff or request.user.is_support:
             return queryset
+
         return queryset.filter(user=request.user)