+++ nodeconductor/structure/models.py
@@ -245,9 +245,11 @@ class PermissionMixin(object):
 
 class CustomerRole(models.CharField):
     OWNER = 'owner'
+    SUPPORT = 'support'
 
     CHOICES = (
         (OWNER, 'Owner'),
+        (SUPPORT, 'Support'),
     )
 
     def __init__(self, *args, **kwargs):
@@ -366,7 +368,14 @@ class Customer(core_models.UuidMixin,
         return get_user_model().objects.filter(
             customerpermission__customer=self,
             customerpermission__is_active=True,
-            customerpermission__role=CustomerRole.OWNER
+            customerpermission__role=CustomerRole.OWNER,
+        )
+
+    def get_support_users(self):
+        return get_user_model().objects.filter(
+            customerpermission__customer=self,
+            customerpermission__is_active=True,
+            customerpermission__role=CustomerRole.SUPPORT,
         )
 
     def get_users(self):
@@ -413,10 +422,12 @@ class BalanceHistory(models.Model):
 class ProjectRole(models.CharField):
     ADMINISTRATOR = 'admin'
     MANAGER = 'manager'
+    SUPPORT = 'support'
 
     CHOICES = (
         (ADMINISTRATOR, 'Administrator'),
         (MANAGER, 'Manager'),
+        (SUPPORT, 'Support'),
     )
 
     def __init__(self, *args, **kwargs):
@@ -494,13 +505,14 @@ class Project(core_models.DescribableMixin,
         return self.name
 
     def get_users(self, role=None):
-        users = get_user_model().objects.filter(
+        query = Q(
             projectpermission__project=self,
             projectpermission__is_active=True,
         )
         if role:
-            users = users.filter(projectpermission__role=role)
-        return users
+            query = query & Q(projectpermission__role=role)
+
+        return get_user_model().objects.filter(query)
 
     def __str__(self):
         return '%(name)s | %(customer)s' % {