+++ nodeconductor/structure/tests/test_customer.py
@@ -119,140 +119,81 @@ class CustomerRoleTest(TransactionTestCase):
         self.assertEqual(0, self.customer.get_owners().count())
 
 
+@ddt
 class CustomerApiPermissionTest(UrlResolverMixin, test.APITransactionTestCase):
     def setUp(self):
-        self.users = {
-            'staff': factories.UserFactory(is_staff=True),
-            'owner': factories.UserFactory(),
-            'not_owner': factories.UserFactory(),
-            'admin': factories.UserFactory(),
-            'admin_other': factories.UserFactory(),
-            'manager': factories.UserFactory(),
-        }
-
-        self.customers = {
-            'owned': factories.CustomerFactory.create_batch(2),
-            'inaccessible': factories.CustomerFactory.create_batch(2),
-            'admin': factories.CustomerFactory(),
-            'manager': factories.CustomerFactory(),
-        }
-
-        for customer in self.customers['owned']:
-            customer.add_user(self.users['owner'], CustomerRole.OWNER)
-
-        self.projects = {
-            'admin': factories.ProjectFactory(customer=self.customers['admin']),
-            'manager': factories.ProjectFactory(customer=self.customers['manager']),
-        }
-
-        self.projects['admin'].add_user(self.users['admin'], ProjectRole.ADMINISTRATOR)
-        self.projects['manager'].add_user(self.users['manager'], ProjectRole.MANAGER)
+        self.fixture = fixtures.ProjectFixture()
 
     # List filtration tests
-    def test_user_can_list_customers_he_is_owner_of(self):
-        self.client.force_authenticate(user=self.users['owner'])
+    @data('staff', 'global_support', 'owner', 'customer_support', 'admin', 'manager', 'project_support')
+    def test_user_can_list_customers(self, user):
+        self.client.force_authenticate(user=getattr(self.fixture, user))
 
-        self._check_user_list_access_customers(self.customers['owned'], 'assertIn')
+        self._check_user_list_access_customers(self.fixture.customer, 'assertIn')
 
-    def test_user_cannot_list_customers_he_is_not_owner_of(self):
-        self.client.force_authenticate(user=self.users['not_owner'])
-
-        self._check_user_list_access_customers(self.customers['inaccessible'], 'assertNotIn')
-
-    def test_user_can_list_customers_if_he_is_staff(self):
-        self.client.force_authenticate(user=self.users['staff'])
-
-        self._check_user_list_access_customers(self.customers['owned'], 'assertIn')
-
-        self._check_user_list_access_customers(self.customers['inaccessible'], 'assertIn')
-
-    def test_user_can_list_customer_if_he_is_admin_in_a_project_owned_by_a_customer(self):
-        self.client.force_authenticate(user=self.users['admin'])
-        self._check_customer_in_list(self.customers['admin'])
-
-    def test_user_can_list_customer_if_he_is_manager_in_a_project_owned_by_a_customer(self):
-        self.client.force_authenticate(user=self.users['manager'])
-        self._check_customer_in_list(self.customers['manager'])
-
-    def test_user_cannot_list_customer_if_he_is_admin_in_a_project_not_owned_by_a_customer(self):
-        self.client.force_authenticate(user=self.users['admin'])
-        self._check_customer_in_list(self.customers['manager'], False)
-
-    def test_user_cannot_list_customer_if_he_is_manager_in_a_project_not_owned_by_a_customer(self):
-        self.client.force_authenticate(user=self.users['manager'])
-        self._check_customer_in_list(self.customers['admin'], False)
+    @data('user', 'admin', 'manager', 'project_support')
+    def test_user_cannot_list_other_customer(self, user):
+        customer = factories.CustomerFactory()
+        self.client.force_authenticate(user=getattr(self.fixture, user))
+        self._check_customer_in_list(customer, False)
 
     # Nested objects filtration tests
-    def test_user_can_see_project_he_has_a_role_in_within_customer(self):
-        for user_role in ('admin', 'manager'):
-            self.client.force_authenticate(user=self.users[user_role])
+    @data('admin', 'manager', 'project_support')
+    def test_user_can_see_project_he_has_a_role_in_within_customer(self, user):
+        self.client.force_authenticate(user=getattr(self.fixture, user))
 
-            customer = self.customers[user_role]
+        response = self.client.get(self._get_customer_url(self.fixture.customer))
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
 
-            seen_project = self.projects[user_role]
+        project_urls = set([project['url'] for project in response.data['projects']])
+        self.assertIn(
+            self._get_project_url(self.fixture.project), project_urls,
+            'User should see project',
+        )
 
-            response = self.client.get(self._get_customer_url(customer))
-            self.assertEqual(response.status_code, status.HTTP_200_OK)
+    @data('admin', 'manager', 'project_support')
+    def test_user_cannot_see_project_he_has_no_role_in_within_customer(self, user):
+        self.client.force_authenticate(user=getattr(self.fixture, user))
 
-            project_urls = set([project['url'] for project in response.data['projects']])
-            self.assertIn(
-                self._get_project_url(seen_project), project_urls,
-                'User should see project',
-            )
+        non_seen_project = factories.ProjectFactory(customer=self.fixture.customer)
 
-    def test_user_cannot_see_project_he_has_no_role_in_within_customer(self):
-        for user_role in ('admin', 'manager'):
-            self.client.force_authenticate(user=self.users[user_role])
+        response = self.client.get(self._get_customer_url(self.fixture.customer))
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
 
-            customer = self.customers[user_role]
-
-            non_seen_project = factories.ProjectFactory(customer=customer)
-
-            response = self.client.get(self._get_customer_url(customer))
-            self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-            project_urls = set([project['url'] for project in response.data['projects']])
-            self.assertNotIn(
-                self._get_project_url(non_seen_project), project_urls,
-                'User should not see project',
-            )
+        project_urls = set([project['url'] for project in response.data['projects']])
+        self.assertNotIn(
+            self._get_project_url(non_seen_project), project_urls,
+            'User should not see project',
+        )
 
     # Direct instance access tests
-    def test_user_can_access_customers_he_is_owner_of(self):
-        self.client.force_authenticate(user=self.users['owner'])
+    @data(('owner', 'owners'), ('customer_support', 'support_users'))
+    def test_user_can_see_its_owner_membership_in_a_service_he_is_owner_of(self, user_data):
+        user, response_field = user_data
+        self.client.force_authenticate(user=getattr(self.fixture, user))
+        response = self.client.get(self._get_customer_url(self.fixture.customer))
+        users = set(c['url'] for c in response.data[response_field])
 
-        self._check_user_direct_access_customer(self.customers['owned'], status.HTTP_200_OK)
+        user_url = self._get_user_url(getattr(self.fixture, user))
+        self.assertItemsEqual([user_url], users)
 
-    def test_user_can_see_its_owner_membership_in_a_service_he_is_owner_of(self):
-        self.client.force_authenticate(user=self.users['owner'])
-        for customer in self.customers['owned']:
-            response = self.client.get(self._get_customer_url(customer))
-            owners = set(c['url'] for c in response.data['owners'])
-            user_url = self._get_user_url(self.users['owner'])
-            self.assertItemsEqual([user_url], owners)
+    @data('staff', 'global_support')
+    def test_user_can_access_all_customers_if_he_is_staff(self, user):
+        self.client.force_authenticate(user=getattr(self.fixture, user))
 
-    def test_user_cannot_access_customers_he_is_not_owner_of(self):
-        self.client.force_authenticate(user=self.users['not_owner'])
-        # 404 is used instead of 403 to hide the fact that the resource exists at all
-        self._check_user_direct_access_customer(self.customers['inaccessible'], status.HTTP_404_NOT_FOUND)
+        self._check_user_direct_access_customer(self.fixture.customer, status.HTTP_200_OK)
 
-    def test_user_can_access_all_customers_if_he_is_staff(self):
-        self.client.force_authenticate(user=self.users['staff'])
-
-        self._check_user_direct_access_customer(self.customers['owned'], status.HTTP_200_OK)
-
-        self._check_user_direct_access_customer(self.customers['inaccessible'], status.HTTP_200_OK)
+        customer = factories.CustomerFactory()
+        self._check_user_direct_access_customer(customer, status.HTTP_200_OK)
 
     # Helper methods
-    def _check_user_list_access_customers(self, customers, test_function):
+    def _check_user_list_access_customers(self, customer, test_function):
         response = self.client.get(reverse('customer-list'))
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
         urls = set([instance['url'] for instance in response.data])
-        for customer in customers:
-            url = self._get_customer_url(customer)
-
-            getattr(self, test_function)(url, urls)
+        url = self._get_customer_url(customer)
+        getattr(self, test_function)(url, urls)
 
     def _check_customer_in_list(self, customer, positive=True):
         response = self.client.get(reverse('customer-list'))
@@ -265,126 +206,98 @@ class CustomerApiPermissionTest(UrlResolverMixin, test.APITransactionTestCase):
         else:
             self.assertNotIn(customer_url, urls)
 
-    def _check_user_direct_access_customer(self, customers, status_code):
-        for customer in customers:
-            response = self.client.get(self._get_customer_url(customer))
-
-            self.assertEqual(response.status_code, status_code)
+    def _check_user_direct_access_customer(self, customer, status_code):
+        response = self.client.get(self._get_customer_url(customer))
+        self.assertEqual(response.status_code, status_code)
 
 
+@ddt
 class CustomerApiManipulationTest(UrlResolverMixin, test.APISimpleTestCase):
     def setUp(self):
-        self.users = {
-            'staff': factories.UserFactory(is_staff=True),
-            'owner': factories.UserFactory(),
-            'not_owner': factories.UserFactory(),
-        }
-
-        self.customers = {
-            'owner': factories.CustomerFactory(),
-            'inaccessible': factories.CustomerFactory(),
-        }
-
-        self.customers['owner'].add_user(self.users['owner'], CustomerRole.OWNER)
+        self.fixture = fixtures.ProjectFixture()
 
     # Deletion tests
-    def test_user_cannot_delete_customer_he_is_not_owner_of(self):
-        self.client.force_authenticate(user=self.users['not_owner'])
+    @data('owner', 'admin', 'manager', 'global_support', 'customer_support', 'project_support')
+    def test_user_cannot_delete_customer(self, user):
+        self.client.force_authenticate(user=getattr(self.fixture, user))
 
-        response = self.client.delete(self._get_customer_url(self.customers['inaccessible']))
-
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_user_cannot_delete_customer_he_is_owner_of(self):
-        self.client.force_authenticate(user=self.users['owner'])
-
-        response = self.client.delete(self._get_customer_url(self.customers['owner']))
-
+        response = self.client.delete(self._get_customer_url(self.fixture.customer))
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_user_cannot_delete_customer_with_associated_projects_if_he_is_staff(self):
-        self.client.force_authenticate(user=self.users['staff'])
+        self.client.force_authenticate(user=self.fixture.staff)
 
-        for customer in self.customers.values():
-            factories.ProjectFactory(customer=customer)
+        factories.ProjectFactory(customer=self.fixture.customer)
+        response = self.client.delete(self._get_customer_url(self.fixture.customer))
 
-            response = self.client.delete(self._get_customer_url(customer))
-
-            self.assertEqual(response.status_code, status.HTTP_409_CONFLICT)
-            self.assertDictContainsSubset({'detail': 'Cannot delete customer with existing projects'},
-                                          response.data)
+        self.assertEqual(response.status_code, status.HTTP_409_CONFLICT)
+        self.assertDictContainsSubset({'detail': 'Cannot delete customer with existing projects'},
+                                      response.data)
 
     # Creation tests
-    def test_user_can_not_create_customer_if_he_is_not_staff(self):
-        self.client.force_authenticate(user=self.users['not_owner'])
+    @data('user', 'global_support')
+    def test_user_can_not_create_customer_if_he_is_not_staff(self, user):
+        self.client.force_authenticate(user=getattr(self.fixture, user))
 
-        response = self.client.post(reverse('customer-list'), self._get_valid_payload())
-
+        response = self.client.post(factories.CustomerFactory.get_list_url(), self._get_valid_payload())
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_user_can_create_customer_if_he_is_not_staff_if_settings_are_tweaked(self):
         add_permission_logic(Customer, OWNER_CAN_MANAGE_CUSTOMER_LOGICS)
-        self.client.force_authenticate(user=self.users['not_owner'])
-
-        response = self.client.post(reverse('customer-list'), self._get_valid_payload())
+        self.client.force_authenticate(user=self.fixture.user)
+        response = self.client.post(factories.CustomerFactory.get_list_url(), self._get_valid_payload())
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
 
         # User became owner of created customer
-        self.assertEqual(response.data['owners'][0]['uuid'], self.users['not_owner'].uuid.hex)
+        self.assertEqual(response.data['owners'][0]['uuid'], self.fixture.user.uuid.hex)
         remove_permission_logic(Customer, OWNER_CAN_MANAGE_CUSTOMER_LOGICS)
 
     def test_user_can_create_customer_if_he_is_staff(self):
-        self.client.force_authenticate(user=self.users['staff'])
+        self.client.force_authenticate(user=self.fixture.staff)
 
-        response = self.client.post(reverse('customer-list'), self._get_valid_payload())
+        response = self.client.post(factories.CustomerFactory.get_list_url(), self._get_valid_payload())
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
 
     # Mutation tests
-    def test_user_cannot_change_customer_as_whole_he_is_not_owner_of(self):
-        self.client.force_authenticate(user=self.users['not_owner'])
+    @data('manager', 'admin', 'customer_support', 'project_support', 'global_support')
+    def test_user_cannot_change_customer_as_whole(self, user):
+        self.client.force_authenticate(user=getattr(self.fixture, user))
 
-        response = self.client.put(self._get_customer_url(self.customers['inaccessible']),
+        response = self.client.put(self._get_customer_url(self.fixture.customer),
                                    self._get_valid_payload())
 
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_user_cannot_change_customer_he_is_owner_of(self):
-        self.client.force_authenticate(user=self.users['owner'])
+        self.client.force_authenticate(user=self.fixture.owner)
 
-        response = self.client.put(self._get_customer_url(self.customers['owner']),
+        response = self.client.put(self._get_customer_url(self.fixture.customer),
                                    self._get_valid_payload())
 
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_user_can_change_customer_as_whole_if_he_is_staff(self):
-        self.client.force_authenticate(user=self.users['staff'])
+        self.client.force_authenticate(user=self.fixture.staff)
 
-        response = self.client.put(self._get_customer_url(self.customers['owner']),
+        response = self.client.put(self._get_customer_url(self.fixture.customer),
                                    self._get_valid_payload())
         self.assertEqual(response.status_code, status.HTTP_200_OK, 'Error message: %s' % response.data)
 
-        response = self.client.put(self._get_customer_url(self.customers['inaccessible']),
-                                   self._get_valid_payload())
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-
     def test_user_cannot_change_single_customer_field_he_is_not_owner_of(self):
-        self.client.force_authenticate(user=self.users['not_owner'])
+        self.client.force_authenticate(user=self.fixture.user)
 
-        self._check_single_customer_field_change_permission(self.customers['inaccessible'], status.HTTP_404_NOT_FOUND)
+        self._check_single_customer_field_change_permission(self.fixture.customer, status.HTTP_404_NOT_FOUND)
 
     def test_user_cannot_change_customer_field_he_is_owner_of(self):
-        self.client.force_authenticate(user=self.users['owner'])
+        self.client.force_authenticate(user=self.fixture.owner)
 
-        self._check_single_customer_field_change_permission(self.customers['owner'], status.HTTP_403_FORBIDDEN)
+        self._check_single_customer_field_change_permission(self.fixture.customer, status.HTTP_403_FORBIDDEN)
 
     def test_user_can_change_single_customer_field_if_he_is_staff(self):
-        self.client.force_authenticate(user=self.users['staff'])
-
-        self._check_single_customer_field_change_permission(self.customers['owner'], status.HTTP_200_OK)
-
-        self._check_single_customer_field_change_permission(self.customers['inaccessible'], status.HTTP_200_OK)
+        self.client.force_authenticate(user=self.fixture.staff)
+        self._check_single_customer_field_change_permission(self.fixture.customer, status.HTTP_200_OK)
 
     # Helper methods
     def _get_valid_payload(self, resource=None):
@@ -529,25 +442,30 @@ class CustomerUnicodeTest(TransactionTestCase):
 class CustomerUsersListTest(test.APITransactionTestCase):
     def setUp(self):
         self.fixture = fixtures.ProjectFixture()
+        self.staff = self.fixture.staff
         self.owner = self.fixture.owner
         self.admin = self.fixture.admin
         self.manager = self.fixture.manager
         self.customer = self.fixture.customer
+        self.customer_support = self.fixture.customer_support
+        self.project_support = self.fixture.project_support
+        self.global_support = self.fixture.global_support
         self.url = factories.CustomerFactory.get_url(self.customer, action='users')
 
-    @data('staff', 'owner', 'manager', 'admin')
+    @data('staff', 'owner', 'manager', 'admin', 'customer_support', 'project_support', 'global_support')
     def test_user_can_list_customer_users(self, user):
-        self.client.force_authenticate(getattr(self.fixture, user))
+        self.client.force_authenticate(getattr(self, user))
         response = self.client.get(self.url)
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data), 3)
+        self.assertEqual(len(response.data), 5)
 
-        self.assertSetEqual({user['role'] for user in response.data}, {'owner', None, None})
+        self.assertSetEqual({user['role'] for user in response.data}, {'owner', 'support', None, None})
         self.assertSetEqual({user['uuid'] for user in response.data},
-                            {self.owner.uuid.hex, self.admin.uuid.hex, self.manager.uuid.hex})
+                            {self.owner.uuid.hex, self.admin.uuid.hex, self.manager.uuid.hex,
+                             self.customer_support.uuid.hex, self.project_support.uuid.hex})
         self.assertSetEqual({user['projects'] and user['projects'][0]['role'] or None
-                             for user in response.data}, {None, 'admin', 'manager'})
+                             for user in response.data}, {None, 'admin', 'manager', 'support'})
 
     def test_user_can_not_list_project_users(self):
         self.client.force_authenticate(self.fixture.user)
@@ -555,21 +473,25 @@ class CustomerUsersListTest(test.APITransactionTestCase):
         self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
 
 
+@ddt
 class CustomerCountersListTest(test.APITransactionTestCase):
     def setUp(self):
         self.fixture = fixtures.ServiceFixture()
         self.owner = self.fixture.owner
+        self.customer_support = self.fixture.customer_support
         self.admin = self.fixture.admin
         self.manager = self.fixture.manager
+        self.project_support = self.fixture.project_support
         self.customer = self.fixture.customer
         self.service = self.fixture.service
         self.url = factories.CustomerFactory.get_url(self.customer, action='counters')
 
-    def test_user_can_get_customer_counters(self):
-        self.client.force_authenticate(self.fixture.owner)
+    @data('owner', 'customer_support')
+    def test_user_can_get_customer_counters(self, user):
+        self.client.force_authenticate(getattr(self.fixture, user))
         response = self.client.get(self.url, {'fields': ['users', 'projects', 'services']})
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data, {'users': 3, 'projects': 1, 'services': 1})
+        self.assertEqual(response.data, {'users': 5, 'projects': 1, 'services': 1})
 
 
 class UserCustomersFilterTest(test.APITransactionTestCase):