+++ nodeconductor/structure/tests/test_project_permissions.py
@@ -40,7 +40,7 @@ class ProjectPermissionSerializerTest(unittest.TestCase):
 
     def test_payload_has_required_fields(self):
         expected_fields = [
-            'url', 'role', 'project', 'project_name', 'pk', 'project_uuid',
+            'url', 'role', 'project', 'project_name', 'pk', 'project_uuid', 'customer_name',
             'created', 'expiration_time', 'created_by',
             'user', 'user_full_name', 'user_native_name', 'user_username', 'user_uuid', 'user_email'
         ]
@@ -506,3 +506,24 @@ class ProjectPermissionCreatedByTest(test.APISimpleTestCase):
 
         permission = ProjectPermission.objects.get(pk=response.data['pk'])
         self.assertEqual(permission.created_by, staff_user)
+
+
+class GetProjectUsersTest(test.APITransactionTestCase):
+
+    def setUp(self):
+        fixture = fixtures.ProjectFixture()
+        self.project = fixture.project
+        self.admin = fixture.admin
+        self.manager = fixture.manager
+
+    def test_get_users_by_default_returns_both_managers_and_admins(self):
+        users = list(self.project.get_users())
+        self.assertListEqual(users, [self.admin, self.manager])
+
+    def test_get_users_by_returns_admins(self):
+        users = list(self.project.get_users(ProjectRole.ADMINISTRATOR))
+        self.assertListEqual(users, [self.admin])
+
+    def test_get_users_by_returns_managers(self):
+        users = list(self.project.get_users(ProjectRole.MANAGER))
+        self.assertListEqual(users, [self.manager])